+++ src/main/java/dev/hoyin1600p/vault_render_optimization/compat/flywheelshader/accessors/BufferBuilderAccessor.java
@@ -1,0 +1,5 @@
+package dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.accessors;
+
+public interface BufferBuilderAccessor {
+    void vroFlywheel$setForceBlockFormat(boolean isFlyWheel);
+}