+++ src/main/java/dev/hoyin1600p/vault_render_optimization/compat/flywheelshader/accessors/ProgramDirectivesAccessor.java
@@ -1,0 +1,7 @@
+package dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.accessors;
+
+import net.coderbot.iris.gl.blending.AlphaTest;
+
+public interface ProgramDirectivesAccessor {
+    void setFlwAlphaTestOverride(AlphaTest alphaTest);
+}