+++ src/main/java/dev/hoyin1600p/vault_render_optimization/compat/flywheelshader/compiler/IrisProgramCompilerBase.java
@@ -1,0 +1,125 @@
+package dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.compiler;
+
+import com.jozufozu.flywheel.backend.gl.shader.GlProgram;
+import com.jozufozu.flywheel.core.compile.ProgramContext;
+import com.jozufozu.flywheel.core.compile.Template;
+import com.jozufozu.flywheel.core.compile.VertexData;
+import com.jozufozu.flywheel.core.shader.WorldProgram;
+import com.jozufozu.flywheel.core.source.FileResolution;
+import net.coderbot.iris.Iris;
+import net.coderbot.iris.gl.blending.AlphaTest;
+import net.coderbot.iris.gl.blending.BlendModeOverride;
+import net.coderbot.iris.pipeline.WorldRenderingPipeline;
+import net.coderbot.iris.pipeline.newshader.FogMode;
+import net.coderbot.iris.shaderpack.ProgramSet;
+import net.coderbot.iris.shaderpack.ProgramSource;
+import net.coderbot.iris.shaderpack.ShaderProperties;
+import net.coderbot.iris.shaderpack.loading.ProgramId;
+import net.coderbot.iris.vertices.IrisVertexFormats;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.ShaderInstance;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.resources.ResourceLocation;
+import org.jetbrains.annotations.NotNull;
+import dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.VroFlywheelShaderCompat;
+import dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.FlywheelShaderCompatState;
+import dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.accessors.IrisRenderingPipelineAccessor;
+import dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.accessors.ProgramSourceAccessor;
+import dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.accessors.WorldProgramAccessor;
+import dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.flywheel.IrisFlwCompatShaderWarp;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public abstract class IrisProgramCompilerBase<P extends WorldProgram> {
+    Map<WorldRenderingPipeline, HashMap<ProgramContext, P>> programCache = new HashMap<>();
+
+    Map<WorldRenderingPipeline, HashMap<ProgramContext, P>> shadowProgramCache = new HashMap<>();
+
+    protected final GlProgram.Factory<P> factory;
+    private static int programCounter = 0;
+
+    public IrisProgramCompilerBase(GlProgram.Factory<P> factory, Template<? extends VertexData> ignoredTemplate, FileResolution ignoredHeader) {
+        this.factory = factory;
+    }
+
+    public P getProgram(ProgramContext ctx, boolean isShadow) {
+
+        if (FlywheelShaderCompatState.isRenderPathActive()) {
+            WorldRenderingPipeline pipeline = Iris.getPipelineManager().getPipelineNullable();
+            HashMap<ProgramContext, P> cache;
+            if (isShadow) {
+                cache = shadowProgramCache.computeIfAbsent(pipeline, key -> new HashMap<>());
+            } else {
+                cache = programCache.computeIfAbsent(pipeline, key -> new HashMap<>());
+            }
+            if (!cache.containsKey(ctx)) {
+                P created = createIrisShaderProgram(ctx, isShadow);
+                cache.put(ctx, created);
+                if (created == null) {
+                    FlywheelShaderCompatState.recordFailure(ctx.spec.name.toString(), null);
+                } else {
+                    FlywheelShaderCompatState.recordSuccess();
+                }
+            }
+            return cache.get(ctx);
+        }
+        return null;
+    }
+
+    private String getFlwShaderName(ResourceLocation location, boolean isShadow) {
+        String randomId = String.valueOf(programCounter);
+        programCounter++;
+        if (isShadow)
+            return String.format("shadow_flw_%s_%s_%s", location.getNamespace(),
+                    location.getPath(), randomId);
+        else
+            return String.format("gbuffers_flw_%s_%s_%s", location.getNamespace(),
+                    location.getPath(), randomId);
+    }
+
+    abstract P createIrisShaderProgram(ProgramContext ctx, boolean isShadow);
+
+    protected P createWorldProgramBySource(ProgramContext ctx, boolean isShadow, IrisRenderingPipelineAccessor pipeline, ProgramSource processedSource) {
+        ShaderInstance override = null;
+        try {
+            if (isShadow) {
+                override = pipeline.callCreateShadowShader(
+                        getFlwShaderName(ctx.spec.name, true), processedSource, ProgramId.Block, AlphaTest.ALWAYS,
+                        IrisVertexFormats.TERRAIN, false, false, false);
+            } else {
+                override = pipeline.callCreateShader(
+                        getFlwShaderName(ctx.spec.name, false), processedSource, ProgramId.Block, AlphaTest.ALWAYS,
+                        IrisVertexFormats.TERRAIN, FogMode.OFF, false, false, false, false);
+            }
+
+        } catch (Exception exception) {
+            VroFlywheelShaderCompat.LOGGER.error("Fail to compile shader", exception);
+            FlywheelShaderCompatState.recordFailure(ctx.spec.name.toString(), exception);
+        }
+
+        if (override != null) {
+            P program = factory.create(ctx.spec.name, override.getId());
+            ((WorldProgramAccessor) program).setShader(new IrisFlwCompatShaderWarp(override));
+            return program;
+        }
+        return null;
+    }
+
+    @NotNull
+    protected ProgramSource programSourceOverrideVertexSource(ProgramContext ctx, ProgramSet programSet, ProgramSource source, String vertexSource) {
+        ShaderProperties properties = ((ProgramSourceAccessor) source).getShaderProperties();
+        BlendModeOverride blendModeOverride = ((ProgramSourceAccessor) source).getBlendModeOverride();
+        //Get a copy of program
+        return new ProgramSource(source.getName() + "_" + ctx.spec.name.getNamespace() + "_" +
+                ctx.spec.name.getPath(), vertexSource,
+                source.getGeometrySource().orElse(null),
+                source.getFragmentSource().orElse(null), programSet, properties, blendModeOverride);
+    }
+
+    public void clear() {
+        programCache.clear();
+        shadowProgramCache.clear();
+    }
+}