+++ src/main/java/dev/hoyin1600p/vault_render_optimization/compat/flywheelshader/vertex/ExtendedLayoutItems.java
@@ -1,0 +1,8 @@
+package dev.hoyin1600p.vault_render_optimization.compat.flywheelshader.vertex;
+
+import com.jozufozu.flywheel.backend.gl.GlNumericType;
+import com.jozufozu.flywheel.core.layout.PrimitiveItem;
+
+public class ExtendedLayoutItems {
+    public static final PrimitiveItem SHORT2 = new PrimitiveItem(GlNumericType.SHORT, 2);
+}